refactor: split interactive_structure_editor.ts into focused mixins (TB-ARCH-2)
Extract marquee rubber-band selection into a new MarqueeSelectionMixin
and group-transform pivot/centroid math plus the group-only halves of
the TransformControls change/mouseUp listeners into a new
GroupTransformMixin, composed alongside InteractiveStructureEditorMixin
in wave.js. The direct-drag lifecycle (beginAtomDrag_/endAtomDrag_/
cancelAtomDrag_) stays in the base mixin - it's a genuinely unified
state machine covering all 4 drag shapes at once, and this round's
Esc-cancel regression tests are concentrated exactly there, so forcing
a single/group split there would trade real correctness coverage for
a readability preference.

Base file: 1460 -> 1211 lines. Pure refactor: all 181 pre-existing
tests pass unchanged, and behavioral equivalence was further verified
live against a real (non-jsdom) WebGL Wave instance - marquee-select,
group direct-drag, and gizmo group-rotate all independently confirmed
correct.
