+++ src/wave.js
@@ -10,9 +10,11 @@ import { BondsMixin } from "./mixins/bonds";
 import { BoundaryMixin } from "./mixins/boundary";
 import { CellMixin } from "./mixins/cell";
 import { ControlsMixin } from "./mixins/controls";
+import { GroupTransformMixin } from "./mixins/group_transform";
 import { ImageMixin } from "./mixins/image";
 import { InteractiveStructureEditorMixin } from "./mixins/interactive_structure_editor";
 import { AllLabelsMixin } from "./mixins/labels/all";
+import { MarqueeSelectionMixin } from "./mixins/marquee_selection";
 import { AllMeasurementsMixin } from "./mixins/measurements/all";
 import { RepetitionMixin } from "./mixins/repetition";
 import SETTINGS from "./settings";
@@ -259,6 +261,8 @@ export class Wave extends mix(WaveBase).with(
     AllLabelsMixin,
     AllMeasurementsMixin,
     ImageMixin,
+    MarqueeSelectionMixin,
+    GroupTransformMixin,
     InteractiveStructureEditorMixin,
 ) {
     /**
